Bugfix where Seeding was pulling an error.

+++ app/models/trophy.rb
@@ -3,6 +3,7 @@ class Trophy < ActiveRecord::Base
   belongs_to :user
 
   def checker(score)
+    score = score.to_s
     @game = self.game_id
     @user = self.user_id
     @theGame = Game.find(@game)
@@ -43,37 +44,40 @@ class Trophy < ActiveRecord::Base
   end
 
   def allTime(score)
-    perfectScore = @theGame.detailscore.last
-    history = UserGameLog.where(game_id: @game).order(:score)
-    highscore = perfectScore == 0 ? history.first.score : history.reverse.first.score
-    top_score_list = []
-    if score.to_i > highscore
-      Notification.create(
-                          game_id: @game,
-                          user_id: @user,
-                          message: "You got the new high score in " +
-                                    @theGame.name +
-                                    " of " + score +
-                                    ". You win an additional 100 coins!",
-                          title: "High score!",
-                          icon: 2
-      )
-      Trophy.where(uniq_id: 2, game_id: @game).all.each do |beatenUser|
+    oldHistory = UserGameLog.where(game_id: @game)
+    if oldHistory.length > 0
+      perfectScore = @theGame.detailscore.last
+      history = oldHistory.order(:score)
+      highscore = perfectScore == 0 ? history.first.score : history.reverse.first.score
+      top_score_list = []
+      if score.to_i > highscore
         Notification.create(
                             game_id: @game,
-                            user_id: beatenUser.user_id,
-                            message: "Your score has been beaten in " +
+                            user_id: @user,
+                            message: "You got the new high score in " +
                                       @theGame.name +
-                                      " by " + User.find(@user).username+
-                                      "with a " + score + ".",
-                            title: "High score beaten!",
-                            icon: 4
+                                      " of " + score +
+                                      ". You win an additional 100 coins!",
+                            title: "High score!",
+                            icon: 2
         )
-        beatenUser.destroy
+        Trophy.where(uniq_id: 2, game_id: @game).all.each do |beatenUser|
+          Notification.create(
+                              game_id: @game,
+                              user_id: beatenUser.user_id,
+                              message: "Your score has been beaten in " +
+                                        @theGame.name +
+                                        " by " + User.find(@user).username+
+                                        "with a " + score + ".",
+                              title: "High score beaten!",
+                              icon: 4
+          )
+          beatenUser.destroy
+        end
+        self.update_attribute(:uniq_id, 2)
+        self.save!
+        @theUser.update_attribute(:coinTo, 100 + @theUser.coinTo)
       end
-      self.update_attribute(:uniq_id, 2)
-      self.save!
-      @theUser.update_attribute(:coinTo, 100 + @theUser.coinTo)
     end
   end
